Squashed even more bugs added feed rate control honed hdf5. Cut myself on honed edge. back to beginning now :/

+++ droogCNC.py
@@ -30,6 +30,10 @@ class TwoAxisStage:
         self.tempFile = None
         self.temprunning = False
         self.filename = None
+        self.buttonx = 6
+        self.buttony = 3
+        self.feedrate = 0
+        self.dro_text = 'X: %1.3f, Y:%1.3f, Feedrate: %d' % (self.pos[0], self.pos[1], self.feedrate)
         self.parameters = {
             'stepPulseLength'   : [0, None],
             'stepIdleDelay'     : [1, None],
@@ -80,30 +84,33 @@ class TwoAxisStage:
 
         # 10 button
         self.btn_10 = tk.Button(master=self.window, text='10', command=lambda: self.switchRate(10))
-        self.btn_10.grid(row=3, column=0, sticky='nsew')
-        self.btn_10.configure(width=15, height=5)
-
+        self.btn_10.grid(row=3, column=1, sticky='nsew')
+        self.btn_10.configure(width=self.buttonx, height=self.buttony)
         # 1 button
         self.btn_1 = tk.Button(master=self.window, text='1', command=lambda: self.switchRate(1))
-        self.btn_1.grid(row=3, column=1, sticky='nsew')
+        self.btn_1.grid(row=3, column=2, sticky='nsew')
+        self.btn_1.configure(width=self.buttonx, height=self.buttony)
 
         # 0.1 button
         self.btn_01 = tk.Button(master=self.window, text='0.1', command=lambda: self.switchRate(0.1))
-        self.btn_01.grid(row=3, column=2, sticky='nsew')
-
+        self.btn_01.grid(row=3, column=3, sticky='nsew')
+        self.btn_01.configure(width=self.buttonx, height=self.buttony)
         # 0.01 button
         self.btn_001 = tk.Button(master=self.window, text='0.01', command=lambda: self.switchRate(0.01))
-        self.btn_001.grid(row=3, column=3, sticky='nsew')
+        self.btn_001.grid(row=3, column=4, sticky='nsew')
+        self.btn_001.configure(width=self.buttonx, height=self.buttony)
 
         self.btn_0001 = tk.Button(master=self.window, text='0.001', command=lambda: self.switchRate(0.001))
-        self.btn_0001.grid(row=3, column=4, sticky='nsew')
+        self.btn_0001.grid(row=3, column=5, sticky='nsew')
+        self.btn_0001.configure(width=self.buttonx, height=self.buttony)
 
         # DRO frame
         self.lbl_frame = tk.Frame(master=self.window, relief=tk.RAISED, borderwidth=3)
-        self.lbl_pos = tk.Label(master=self.lbl_frame, text='X: %1.3f, Y:%1.3f' % (self.pos[0], self.pos[1]))
-        self.lbl_pos['font'] = font.Font(size=15)
+        self.lbl_pos = tk.Label(master=self.lbl_frame, text='X: %1.3f, Y:%1.3f, Feedrate: %d'
+                                                            % (self.pos[0], self.pos[1], self.feedrate))
+        self.lbl_pos['font'] = font.Font(size=12)
 
-        self.lbl_frame.grid(row=0, column=0, columnspan=self.colLen+1, sticky='')
+        self.lbl_frame.grid(row=0, column=0, columnspan=self.colLen+1, sticky='ew')
         self.lbl_pos.pack()
 
         # gcode input box
@@ -115,6 +122,7 @@ class TwoAxisStage:
                                     command=lambda: self.sendCommand(self.gcode_entry.get().rstrip() + '\n', resetarg=True,
                                                                      entry=self.gcode_entry))
         self.gcode_send.grid(row=4, column=3, sticky='')
+        self.gcode_send.configure(width=self.buttonx, height=self.buttony)
 
         # file input box
         self.file_entry = tk.Entry(master=self.window)
@@ -123,30 +131,40 @@ class TwoAxisStage:
         # file input button
         self.file_input = tk.Button(master=self.window, text='Get File', command=lambda: self.getFile(self.file_entry.get()))
         self.file_input.grid(row=5, column=3)
+        self.file_input.configure(width=self.buttonx, height=self.buttony)
 
         # file run button
         self.file_run = tk.Button(master=self.window, text='Run', command=lambda: self.runFile())
         self.file_run.grid(row=5, column=4)
-
+        self.file_run.configure(width=self.buttonx, height=self.buttony)
         self.kill_btn = tk.Button(master=self.window, text='Kill', command=self.killSwitch)
+
         self.kill_btn.grid(row=5, column=5)
+        self.kill_btn.configure(width=self.buttonx, height=self.buttony)
 
         self.increment_btn = tk.Button(master=self.window, text='G91', command=self.setG91)
         self.increment_btn.grid(row=1, column=1, sticky='nsew')
+        self.increment_btn.configure(width=self.buttonx, height=self.buttony)
 
         self.absolute_btn = tk.Button(master=self.window, text='G90', command=self.setG90)
         self.absolute_btn.grid(row=1, column=3, sticky='nsew')
+        self.absolute_btn.configure(width=self.buttonx, height=self.buttony)
+
         # go home button
-        self.home_btn = tk.Button(master=self.window, text='go\nhome', command=lambda: self.sendCommand('G90 X0 Y0'))
+        self.home_btn = tk.Button(master=self.window, text='Go\nHome', command=lambda: self.sendCommand('G90 X0 Y0 F'
+                                  + str(self.parameters['xMaxRate'][1])))
         self.home_btn.grid(row=2, column=4, sticky='nsew')
+        self.home_btn.configure(width=self.buttonx, height=self.buttony)
 
         # set home button
-        self.set_home = tk.Button(master=self.window, text='set\nhome', command=lambda: self.sendCommand('G92 X0 Y0'))
+        self.set_home = tk.Button(master=self.window, text='Set\nHome', command=lambda: self.sendCommand('G92 X0 Y0'))
         self.set_home.grid(row=1, column=4, sticky='nsew')
+        self.set_home.configure(width=self.buttonx, height=self.buttony)
 
         self.start_from_death_btn = tk.Button(master=self.window, text='No temp\nfile found', command=self.__startFromDeath)
         self.start_from_death_btn.grid(row=2, column=5, sticky='nesw')
-        self.start_from_death_btn['font'] = font.Font(size=8)
+        self.start_from_death_btn['font'] = font.Font(size=7)
+        self.start_from_death_btn.configure(width=self.buttonx, height=self.buttony)
 
         self.__setTempFile()
         self.Refresh()
@@ -163,7 +181,7 @@ class TwoAxisStage:
         Sets recurring event to update GUI every 50ms
         :return: None
         """
-        self.lbl_pos.configure(text='X: %1.3f, Y:%1.3f' % (float(self.pos[0]), float(self.pos[1])))
+        self.lbl_pos.configure(text='X: %1.3f, Y:%1.3f, Feedrate: %d' % (self.pos[0], self.pos[1], self.feedrate))
         self.window.after(50, self.Refresh)
 
     def onKeyPress(self, event, wasd=False):
@@ -219,7 +237,7 @@ class TwoAxisStage:
         """
         # todo: Fix borked DRO feedback
         # implement own method?
-        out = self.s.readline()  # Wait for grbl response with carriage return
+        out = self.s.readline()        # Wait for grbl response with carriage return
         print('> ' + out.strip().decode('UTF-8'))
 
     def sendCommand(self, gcode, resetarg=False, entry=None):
@@ -234,15 +252,18 @@ class TwoAxisStage:
         if self.s:
             if gcode.strip() == '' or gcode.strip()[0] == ';':
                 return
-
+            t = gcode.rstrip().lower().split(' ')
             # check if it's a manual entry to clear entry box
             if resetarg and entry is not None:
                 entry.delete(0, 'end')
 
-            if 'g90' in gcode.rstrip().lower().split(' '):
+            if 'g90' in t:
                 self.setG90(cmd=False)
-            if 'g91' in gcode.rstrip().lower().split(' '):
+            if 'g91' in t:
                 self.setG91(cmd=False)
+            for i in t:
+                if i.lower().strip()[0] == 'f':
+                    self.__setFeed(int(i.strip()[1:]))
             print('Sent: ' + gcode.rstrip())
             gcode = gcode.rstrip() + '\n'
             self.s.write(gcode.encode('UTF-8'))
@@ -282,6 +303,7 @@ class TwoAxisStage:
                 self.__parseParameters()
 
         else:
+            self.connected = False
             self.s.close()
             self.s = None
             self.queue = None
@@ -289,16 +311,18 @@ class TwoAxisStage:
             self.connect_btn.configure(fg='red', text='Connect')
 
     def setG91(self, cmd=True):
-        if cmd:
-            self.sendCommand('G91')
-        self.increment_btn.configure(fg='green')
-        self.absolute_btn.configure(fg='black')
+        if self.s:
+            if cmd:
+                self.sendCommand('G91')
+            self.increment_btn.configure(fg='green')
+            self.absolute_btn.configure(fg='black')
 
     def setG90(self, cmd=True):
-        if cmd:
-            self.sendCommand('G90')
-        self.increment_btn.configure(fg='black')
-        self.absolute_btn.configure(fg='green')
+        if self.s:
+            if cmd:
+                self.sendCommand('G90')
+            self.increment_btn.configure(fg='black')
+            self.absolute_btn.configure(fg='green')
 
     def setPos(self, cmd):
         """
@@ -347,24 +371,25 @@ class TwoAxisStage:
         Used to run a file obained with getFile()
         :return: None
         """
-        if not self.temprunning:
-            print('calling savetemp')
-            self.temprunning = True
-            self.__saveTempData()
+        if self.s:
+            if not self.temprunning:
+                print('calling savetemp')
+                self.temprunning = True
+                self.__saveTempData()
 
-        nextpos = self.queue.dequeue()
-        self.sendCommand(nextpos)
+            nextpos = self.queue.dequeue()
+            self.sendCommand(nextpos)
 
-        if self.queue.size() > 0:
-            self.window.after(self.calcDelay(self.currentpos, nextpos), self.runFile)
-            self.currentpos = nextpos
+            if self.queue.size() > 0:
+                self.window.after(self.calcDelay(self.currentpos, nextpos), self.runFile)
+                self.currentpos = nextpos
 
-        elif self.queue.size() == 0:
-            self.window.after(self.calcDelay(self.currentpos, nextpos), self.finishRun)
+            elif self.queue.size() == 0:
+                self.window.after(self.calcDelay(self.currentpos, nextpos), self.finishRun)
 
-        else:
-            self.temprunning = False
-            return
+            else:
+                self.temprunning = False
+                return
 
     def finishRun(self):
         """
@@ -379,9 +404,9 @@ class TwoAxisStage:
         effectively kills current gcode run by clearing queue. Note this isn't instantaneous
         :return: None
         """
-        self.queue.clear()
-        print('Current motion was killed')
-
+        if self.s:
+            self.queue.clear()
+            print('Current motion was killed')
     def calcDelay(self, currentpos, nextpos):
         """
         Calculates a lower end of the required delay between moved for the queue command system
@@ -437,13 +462,16 @@ class TwoAxisStage:
                 if i.split('=')[0] == str(value[0]):
                     self.parameters[key] = [i.split('=')[0], i.split('=')[1]]
 
+        self.feedrate = int(self.parameters['xMaxRate'][1])
+        print(self.feedrate)
+
     def __setTempFile(self):
         """
         Sets the temporary file name, unless it exists
         :return: None
         """
         if os.path.exists('temp.npy') or self.tempFile is not None:
-            self.start_from_death_btn.configure(text='Load Data?', fg='red')
+            self.start_from_death_btn.configure(text='Load\nData?', fg='red')
             self.__blinkButton(self.start_from_death_btn, 'red', 'blue', 1000)
         else:
             self.tempFile = 'temp.npy'
@@ -468,24 +496,27 @@ class TwoAxisStage:
         return currentline
 
     def __startFromDeath(self):
-        self.queue = None
-        currentline = self.__retrieveTempData()
-        try:
-            f = open(self.filename, 'r')
-            self.queue = Queue()
-            positionFound = False
-            for line in f:
-                if line == currentline:
-                    positionFound = True
-                if positionFound:
-                    if line != '' and line[0] != ';':
-                        self.queue.enqueue(line)
-            print('Loaded ' + self.filename)
-            self.start_from_death_btn.configure(text='Start?', fg='green', command=self.runFile)
-        except FileNotFoundError:
-            self.file_entry.delete(0, 'end')
-            self.file_entry.insert(0, 'File does not exist')
-
+        if self.s:
+            self.queue = None
+            currentline = self.__retrieveTempData()
+            try:
+                f = open(self.filename, 'r')
+                self.queue = Queue()
+                positionFound = False
+                for line in f:
+                    if line == currentline:
+                        positionFound = True
+                    if positionFound:
+                        if line != '' and line[0] != ';':
+                            self.queue.enqueue(line)
+                print('Loaded ' + self.filename)
+                self.start_from_death_btn.configure(text='Start?', fg='green', command=self.runFile)
+            except FileNotFoundError:
+                self.file_entry.delete(0, 'end')
+                self.file_entry.insert(0, 'File does not exist')
+        else:
+            self.start_from_death_btn.configure(text='Not\nconnected')
+            self.window.after(5000, lambda: self.start_from_death_btn.configure(text='Load\nData?'))
     def __removeTempFile(self):
         """
         removes temp file at end of program run
@@ -506,6 +537,9 @@ class TwoAxisStage:
 
     def __writeData(self, data):
         self.datafile.append(data)
+
+    def __setFeed(self, feedrate):
+        self.feedrate = feedrate
 
 
 class Queue:
